added ui for the withdraw page
