+++ frontend/src/components/Navbar.tsx
@@ -79,10 +79,10 @@ const Navbar = () => {
                     <nav className="hidden md:block absolute left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2">
                       <ul className="flex items-center justify-center gap-8">
                         <Link href="/app" className="hover:text-foreground/80 text-sm">
-                          Dashboard
+                          Deposit
                         </Link>
-                        <Link href="stake" className="hover:text-foreground/80 text-sm">
-                          Stake
+                        <Link href="withdraw" className="hover:text-foreground/80 text-sm">
+                          Withdraw
                         </Link>
                         <Link href="#" className="hover:text-foreground/80 text-sm">
                           Lucky Draw
